move chatbot into subfolder
