+++ lib/peppermint_sdk.dart
@@ -15,11 +15,11 @@ export 'src/widgets/media/media_gram.dart';
 export 'src/widgets/nft_scanner.dart';
 export 'src/resource.dart';
 
-export 'src/chatbot/repo/chatbot_repo.dart';
-export 'src/chatbot/usecases/create_conversation_usecase.dart';
-export 'src/chatbot/usecases/get_messages_usecase.dart';
-export 'src/chatbot/usecases/send_message_usecase.dart';
-export 'src/chatbot/models/chat_message_response.dart';
-export 'src/chatbot/models/chatbot_model.dart';
-export 'src/chatbot/data_source/chat_local_data_source.dart';
-export 'src/chatbot/controller/chatbot_controller.dart';
+export 'src/peppermint_functionalities/chatbot/repo/chatbot_repo.dart';
+export 'src/peppermint_functionalities/chatbot/usecases/create_conversation_usecase.dart';
+export 'src/peppermint_functionalities/chatbot/usecases/get_messages_usecase.dart';
+export 'src/peppermint_functionalities/chatbot/usecases/send_message_usecase.dart';
+export 'src/peppermint_functionalities/chatbot/models/chat_message_response.dart';
+export 'src/peppermint_functionalities/chatbot/models/chatbot_model.dart';
+export 'src/peppermint_functionalities/chatbot/data_source/chat_local_data_source.dart';
+export 'src/peppermint_functionalities/chatbot/controller/chatbot_controller.dart';
