+++ lib/src/peppermint_functionalities/chatbot/models/chat_message_response.dart
