+++ lib/src/peppermint_functionalities/chatbot/models/chatbot_model.dart
