+++ lib/src/peppermint_functionalities/chatbot/usecases/get_messages_usecase.dart
@@ -1,4 +1,4 @@
-import '../../resource.dart';
+import '../../../resource.dart';
 import '../models/chat_message_response.dart';
 import '../repo/chatbot_repo.dart';
